+++ e2e/__tests__/filter.test.ts
@@ -43,4 +43,29 @@ describe('Dynamic test filtering', () => {
     expect(result.stderr).toContain('did not return a valid test list');
     expect(result.stderr).toContain('my-clowny-filter');
   });
+
+  it('will call setup on filter before filtering', () => {
+    const result = runJest('filter', ['--filter=<rootDir>/my-setup-filter.js']);
+
+    expect(result.status).toBe(0);
+    expect(result.stderr).toContain('1 total');
+  });
+
+  it('will print error when filter throws', () => {
+    const result = runJest('filter', [
+      '--filter=<rootDir>/my-broken-filter.js',
+    ]);
+
+    expect(result.status).toBe(1);
+    expect(result.stderr).toContain('Error: My broken filter error.');
+  });
+
+  it('will return no results when setup hook throws', () => {
+    const result = runJest('filter', [
+      '--filter=<rootDir>/my-broken-setup-filter.js',
+    ]);
+
+    expect(result.status).toBe(1);
+    expect(result.stderr).toContain('Error: My broken setup filter error.');
+  });
 });